+++ app/tabs/structures.tsx
@@ -1,6 +1,6 @@
 import type { JSX, RefObject, TargetedMouseEvent } from "preact";
 import _React, { render, useEffect, useRef } from "preact/compat";
-import { entryContentTypeToFormatMap, getKeyDisplayName, NBTSchemas, prettyPrintSNBT, prismarineToSNBT } from "mcbe-leveldb";
+import { entryContentTypeToFormatMap, getKeyDisplayName, NBTSchemas, prettyPrintSNBT, prismarineToSNBT, type EntryContentTypeFormatData } from "mcbe-leveldb";
 import NBT from "prismarine-nbt";
 import { existsSync, type Dirent } from "node:fs";
 import path from "node:path";
@@ -17,8 +17,7 @@ import { readdirRecursiveSafe } from "../../src/utils/folderContentsUtils";
 import type { ShowSelectOpenTabDialogResult } from "../components/SelectOpenTabDialog";
 import showSelectOpenTabDialog from "../components/SelectOpenTabDialog";
 import Notice from "../components/Notice";
-
-// TODO: Implement Async Mode for this tab. #57
+import { createObservable, type Observable } from "../../src/utils/miscUtils";
 
 /**
  * Props for the {@link StructuresTab} component.
@@ -319,12 +318,14 @@ export default function StructuresTab(props: StructuresTabProps): JSX.SpecificEl
 interface KeyData {
     rawKey: Buffer;
     displayKey: string;
-    // UNDONE: Uncomment the `?` and ` | undefined` when async mode is implemented.
-    data /* ? */: {
-        parsed: Pick<NBT.NBT, "name"> & NBTSchemas.NBTSchemaTypes.StructureTemplate;
-        type: NBT.NBTFormat;
-        metadata: NBT.Metadata;
-    } | null /* | undefined */;
+    data?:
+        | {
+              parsed: Pick<NBT.NBT, "name"> & NBTSchemas.NBTSchemaTypes.StructureTemplate;
+              type: NBT.NBTFormat;
+              metadata: NBT.Metadata;
+          }
+        | null
+        | undefined;
 }
 
 async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal): Promise<JSX.Element> {
@@ -383,6 +384,12 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
     if (!tab.cachedDBKeys) await tab.awaitCachedDBKeys!;
     signal.throwIfAborted();
     const rawKeys: Buffer[] = tab.cachedDBKeys!.StructureTemplate;
+    const asyncMode: boolean =
+        "__FORCE_ASYNC_KEY_MODE__" in window ? !!window.__FORCE_ASYNC_KEY_MODE__
+        : config.useAsyncModeInEntryViews === "auto" ?
+            rawKeys.length >= config.asyncModeEntryThreshold ||
+            Object.values(tab.cachedDBKeys!).reduce((a: number, b: Buffer[]): number => a + b.length, 0) >= config.asyncModeTotalKeyCountThreshold
+        :   config.useAsyncModeInEntryViews;
     let keys: KeyData[] = await Promise.all(
         rawKeys.map(
             async (key: Buffer): Promise<KeyData> => ({
@@ -392,27 +399,28 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
             })
         )
     );
+    let targetKeys: KeyData[] = keys;
     // globalThis.a = keys;
-    const dynamicProperties: NBT.NBT | undefined = await tab.db
-        .get("DynamicProperties")
-        .then((data: Buffer | null): Promise<NBT.NBT> | undefined =>
-            data ? NBT.parse(data).then((data: { parsed: NBT.NBT; type: NBT.NBTFormat; metadata: NBT.Metadata }): NBT.NBT => data.parsed) : undefined
-        )
-        .catch((e: unknown): undefined => (console.error(e), undefined));
-    // console.log(dynamicProperties);
     let mode: ConfigConstants.views.Structures.StructuresTabMode = config.views.structures.mode;
-    let tablesContents: JSX.Element[][] = await Promise.all(
-        ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode].map(
-            async (sectionID: (typeof ConfigConstants.views.Structures.structuresTabModeToSectionIDs)[typeof mode][number]): Promise<JSX.Element[]> =>
-                await getStructuresTabContentsRows({
-                    tab,
-                    keys,
-                    dynamicProperties,
-                    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions -- There is only one section atm, if another section is ever added, remove this disable comment.
-                    mode: sectionID === null ? mode : `${mode}_${sectionID}`,
-                })
-        )
-    );
+    let currentUpdateTablesContentsFunction: ((reloadData: boolean) => Promise<void>) | null = null;
+    let emptyTablesContents: JSX.Element[][] =
+        asyncMode ?
+            [[]]
+        :   await Promise.all(
+                ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode].map(
+                    async (sectionID: (typeof ConfigConstants.views.Structures.structuresTabModeToSectionIDs)[typeof mode][number]): Promise<JSX.Element[]> =>
+                        await getStructuresTabContentsRows({
+                            tab,
+                            keys,
+                            // eslint-disable-next-line @typescript-eslint/restrict-template-expressions -- There is only one section atm, if another section is ever added, remove this disable comment.
+                            mode: sectionID === null ? mode : `${mode}_${sectionID}`,
+                            get updateTablesContents(): ((reloadData: boolean) => Promise<void>) | null {
+                                return currentUpdateTablesContentsFunction;
+                            },
+                        })
+                )
+            );
+    let tablesContents: JSX.Element[][] = emptyTablesContents;
     function Contents(): JSX.Element {
         const tablesContainerRef: RefObject<HTMLTableElement> = useRef<HTMLTableElement>(null);
         const loadingScreenMessageContainerRef: RefObject<HTMLDivElement> = useRef<HTMLDivElement>(null);
@@ -427,7 +435,61 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
         //     viewOptionsContainer: useRef<HTMLDivElement>(null),
         //     viewOptionsTabbedSelector: useRef<HTMLDivElement>(null),
         // };
+        async function getTablesContentsInRange(sectionIndex: number, start: number, end: number): Promise<JSX.Element[]> {
+            const sectionID: (typeof ConfigConstants.views.Structures.structuresTabModeToSectionIDs)[typeof mode][number] =
+                ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode][sectionIndex]!;
+            return await getStructuresTabContentsRows({
+                tab,
+                // REVIEW // TEST: Make sure this won't crash the tab if an entry with invalid data is present.
+                keys: await Promise.all(
+                    targetKeys
+                        .slice(start, end)
+                        .map(
+                            async (key: KeyData): Promise<KeyData> => ({ ...key, data: (await NBT.parse((await tab.db!.get(key.rawKey))!)) as KeyData["data"] })
+                        )
+                ),
+                // eslint-disable-next-line @typescript-eslint/restrict-template-expressions -- There is only one section atm, if another section is ever added, remove this disable comment.
+                mode: sectionID === null ? mode : `${mode}_${sectionID}`,
+                get updateTablesContents(): ((reloadData: boolean) => Promise<void>) | null {
+                    return currentUpdateTablesContentsFunction;
+                },
+            });
+        }
+        async function _loadTablesContentsInRange(sectionIndex: number, start: number, end: number): Promise<void> {
+            if (!asyncMode) return;
+            // const sectionID: (typeof ConfigConstants.views.Structures.structuresTabModeToSectionIDs)[typeof mode][number] =
+            //     ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode][sectionIndex]!;
+            tablesContents = [...tablesContents];
+            tablesContents[sectionIndex] = [...emptyTablesContents[sectionIndex]!];
+            tablesContents[sectionIndex].splice(start, end - start, ...(await getTablesContentsInRange(sectionIndex, start, end)));
+        }
+        function getSectionEntryCounts(): number[] {
+            return ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode].map(
+                (sectionID: (typeof ConfigConstants.views.Structures.structuresTabModeToSectionIDs)[typeof mode][number]): number => {
+                    switch (sectionID) {
+                        case null:
+                            return targetKeys.length;
+                        default:
+                            return NaN;
+                    }
+                }
+            );
+        }
         function TablesContents(): JSX.Element {
+            const localTablesContents: Observable<JSX.Element[][]> = createObservable([[]]);
+            if (asyncMode) {
+                // TODO: Add an error handler to this.
+                void Promise.all(
+                    ConfigConstants.views.Maps.mapsTabModeToSectionIDs[mode].map(
+                        async (
+                            _sectionID: (typeof ConfigConstants.views.Maps.mapsTabModeToSectionIDs)[typeof mode][number],
+                            index: number
+                        ): Promise<JSX.Element[]> => await getTablesContentsInRange(index, 0, 20)
+                    )
+                ).then((tablesContents: JSX.Element[][]): void => {
+                    localTablesContents.set(tablesContents);
+                });
+            }
             return (
                 <>
                     {...ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode].map(
@@ -437,6 +499,13 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
                         ): JSX.Element => {
                             function Test1(): JSX.Element {
                                 const bodyRef: RefObject<HTMLTableSectionElement> = useRef<HTMLTableSectionElement>(null);
+                                localTablesContents.observe((tablesContents: JSX.Element[][]): void => {
+                                    if (!asyncMode || !bodyRef.current) return;
+                                    // const tempElement: HTMLDivElement = document.createElement("div");
+                                    render(null, bodyRef.current);
+                                    render(<>{...tablesContents[index]!}</>, bodyRef.current /* tempElement */);
+                                    // bodyRef.current.replaceChildren(...tempElement.children);
+                                });
                                 // const [columnHeadersContextMenu_isOpen, columnHeadersContextMenu_setOpen] = useState(false);
                                 // const [columnHeadersContextMenu_anchorPoint, columnHeadersContextMenu_setAnchorPoint] = useState({ x: 0, y: 0 });
                                 const headerName = ConfigConstants.views.Structures.structuresTabModeSectionHeaderNames[mode][index];
@@ -495,13 +564,24 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
                                                 <tr class="table-footer-row-page-navigation">
                                                     <td colSpan={ConfigConstants.views.Structures.structuresTabModeToColumnIDs[sectionMode].length}>
                                                         <PageNavigation
-                                                            totalPages={Math.ceil(tablesContents[index]!.length / 20)}
-                                                            onPageChange={(page: number): void => {
+                                                            totalPages={Math.ceil(getSectionEntryCounts()[index]! / 20)}
+                                                            onPageChange={async (page: number): Promise<void> => {
                                                                 if (!bodyRef.current) return;
+                                                                if (asyncMode) {
+                                                                    localTablesContents.get()[index] = await getTablesContentsInRange(
+                                                                        index,
+                                                                        (page - 1) * 20,
+                                                                        page * 20
+                                                                    );
+                                                                }
                                                                 // let tempElement: HTMLDivElement = document.createElement("div");
                                                                 render(null, bodyRef.current);
                                                                 render(
-                                                                    <>{...tablesContents[index]!.slice((page - 1) * 20, page * 20)}</>,
+                                                                    <>
+                                                                        {...asyncMode ?
+                                                                            localTablesContents.get()[index]!
+                                                                        :   tablesContents[index]!.slice((page - 1) * 20, page * 20)}
+                                                                    </>,
                                                                     bodyRef.current /* tempElement */
                                                                 );
                                                                 // bodyRef.current.replaceChildren(...tempElement.children);
@@ -520,7 +600,7 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
                 </>
             );
         }
-        const query: Omit<TabManagerTab_LevelDBSearchQuery, "searchTargets"> & {
+        const query: Omit<TabManagerTab_LevelDBSearchQuery<true>, "searchTargets"> & {
             searchTargets: {
                 key: Buffer<ArrayBufferLike>;
                 displayKey: string;
@@ -530,11 +610,14 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
                           type: NBT.NBTFormat;
                           metadata: NBT.Metadata;
                       }
+                    | (() => Promise<{ parsed: NBT.NBT; type: NBT.NBTFormat; metadata: NBT.Metadata } | null | undefined>)
                     | null
                     | undefined;
-                valueType: {
-                    readonly type: "NBT";
-                };
+                valueType:
+                    | {
+                          readonly type: "NBT";
+                      }
+                    | Extract<EntryContentTypeFormatData, { type: "custom"; resultType: "JSONNBT" }>;
                 contentType: "StructureTemplate";
                 data: KeyData;
                 searchableContents: string[];
@@ -545,7 +628,11 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
                     ({
                         key: key.rawKey,
                         displayKey: key.displayKey,
-                        value: key.data,
+                        value:
+                            asyncMode ?
+                                async (): Promise<NonNullable<KeyData["data"]>> =>
+                                    (await NBT.parse((await tab.db!.get(key.rawKey))!)) as NonNullable<KeyData["data"]>
+                            :   key.data!,
                         valueType: entryContentTypeToFormatMap.StructureTemplate,
                         contentType: "StructureTemplate",
                         data: key,
@@ -565,16 +652,25 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
                             structureid: key.displayKey.replace(/^structuretemplate_/, ""),
                             // TODO: Uncomment the below line and implement a search query for checking for entries with invalid data.
                             // hasInvalidData: key.data === null,
-                            contents: ((): string => {
-                                if (key.data === null) return "";
-                                try {
-                                    return prettyPrintSNBT(prismarineToSNBT(key.data.parsed), { indent: 0 });
-                                } catch {
-                                    return "";
-                                }
-                            })(),
+                            contents:
+                                asyncMode ?
+                                    async (): Promise<string> => {
+                                        try {
+                                            return prettyPrintSNBT(prismarineToSNBT((await NBT.parse((await tab.db!.get(key.rawKey))!)).parsed), { indent: 0 });
+                                        } catch {
+                                            return "";
+                                        }
+                                    }
+                                :   ((): string => {
+                                        if (key.data === null) return "";
+                                        try {
+                                            return prettyPrintSNBT(prismarineToSNBT(key.data!.parsed), { indent: 0 });
+                                        } catch {
+                                            return "";
+                                        }
+                                    })(),
                         },
-                    }) as const satisfies NonNullable<TabManagerTab_LevelDBSearchQuery["searchTargets"]>[number]
+                    }) as const satisfies NonNullable<TabManagerTab_LevelDBSearchQuery<true>["searchTargets"]>[number]
             ),
         };
         async function updateTablesContents(reloadData: boolean): Promise<void> {
@@ -583,31 +679,63 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
             if (reloadData) {
                 mode = config.views.structures.mode;
                 console.debug(query);
-                tablesContents = await Promise.all(
-                    ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode].map(
-                        async (
-                            sectionID: (typeof ConfigConstants.views.Structures.structuresTabModeToSectionIDs)[typeof mode][number]
-                        ): Promise<JSX.Element[]> =>
-                            await getStructuresTabContentsRows({
-                                tab,
-                                keys:
-                                    Object.keys(query).length > 1 ?
-                                        tab
-                                            .dbSearch!.search(query)
-                                            .toArray()
-                                            .map((key): KeyData => key.originalObject.data)
-                                    :   keys,
-                                dynamicProperties,
-                                // eslint-disable-next-line @typescript-eslint/restrict-template-expressions -- There is only one section atm, if another section is ever added, remove this disable comment.
-                                mode: sectionID === null ? mode : `${mode}_${sectionID}`,
-                            })
-                    )
-                );
+                if (asyncMode) {
+                    targetKeys =
+                        Object.keys(query).length > 1 ?
+                            await (async (): Promise<KeyData[]> => {
+                                const iterator = tab.dbSearch!.searchAsync(query, true);
+                                let i: number = 0;
+                                let t: number = Date.now();
+                                const results: KeyData[] = [];
+                                const formatter = new Intl.NumberFormat();
+                                for await (const value of iterator) {
+                                    i++;
+                                    if (t + 15 < Date.now()) {
+                                        if (loadingScreenMessageContainerRef.current) {
+                                            loadingScreenMessageContainerRef.current.textContent = `Searching LevelDB: ${formatter.format(i)}/${formatter.format(keys.length)} (${formatter.format(results.length)} results)...`;
+                                        }
+                                        signal.throwIfAborted();
+                                        await sleep(5);
+                                        t = Date.now();
+                                    }
+                                    if (!value) continue;
+                                    results.push(value.originalObject.data);
+                                }
+                                return results;
+                            })()
+                        :   keys;
+                } else {
+                    emptyTablesContents = await Promise.all(
+                        ConfigConstants.views.Structures.structuresTabModeToSectionIDs[mode].map(
+                            async (
+                                sectionID: (typeof ConfigConstants.views.Structures.structuresTabModeToSectionIDs)[typeof mode][number]
+                            ): Promise<JSX.Element[]> =>
+                                await getStructuresTabContentsRows({
+                                    tab,
+                                    keys:
+                                        Object.keys(query).length > 1 ?
+                                            tab
+                                                .dbSearch!.search(query)
+                                                .toArray()
+                                                .map((key): KeyData => key.originalObject.data)
+                                        :   keys,
+                                    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions -- There is only one section atm, if another section is ever added, remove this disable comment.
+                                    mode: sectionID === null ? mode : `${mode}_${sectionID}`,
+                                    get updateTablesContents(): ((reloadData: boolean) => Promise<void>) | null {
+                                        return currentUpdateTablesContentsFunction;
+                                    },
+                                })
+                        )
+                    );
+                    tablesContents = emptyTablesContents;
+                }
             }
-            const tempElement: HTMLDivElement = document.createElement("div");
-            render(<TablesContents />, tempElement);
-            tablesContainerRef.current.replaceChildren(...tempElement.children);
+            // const tempElement: HTMLDivElement = document.createElement("div");
+            render(null, tablesContainerRef.current);
+            render(<TablesContents />, tablesContainerRef.current /* tempElement */);
+            // tablesContainerRef.current.replaceChildren(...tempElement.children);
         }
+        currentUpdateTablesContentsFunction = updateTablesContents;
         useEffect((): (() => void) => {
             function onModeChanged(): void {
                 void updateTablesContents(true);
@@ -876,6 +1004,9 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
                                 const refreshBeforeExporting: boolean = false;
                                 if (refreshBeforeExporting) {
                                     await tab.refreshCachedDBKeys();
+                                    const targetKeysIsKeys: boolean = targetKeys === keys;
+                                    void keys;
+                                    // eslint-disable-next-line require-atomic-updates -- This is a false positive.
                                     keys = await Promise.all(
                                         rawKeys.map(
                                             async (key: Buffer): Promise<KeyData> => ({
@@ -885,6 +1016,8 @@ async function getStructuresTabContents(tab: TabManagerTab, signal: AbortSignal)
                                             })
                                         )
                                     );
+                                    void targetKeys;
+                                    if (targetKeysIsKeys) targetKeys = keys;
                                 }
                                 const structureKeys: KeyData[] =
                                     filterExportsBySearchQuery && Object.keys(query).length > 1 ?
@@ -1388,11 +1521,11 @@ async function getStructuresTabContentsRows(data: {
      * The full list of key data to display.
      */
     keys: KeyData[];
-    dynamicProperties?: NBT.NBT | undefined;
     /**
      * The mode of the tab.
      */
     mode: ConfigConstants.views.Structures.StructuresTabSectionMode;
+    get updateTablesContents(): ((reloadData: boolean) => Promise<void>) | null; // TODO
 }): Promise<JSX.Element[]> {
     // const columns = config
     switch (data.mode) {