+++ Changelog.md
@@ -39,6 +39,7 @@
 -   An error message now appears below the search bar on the "View Files" tab when an invalid NBT query is provided, like it does on other tabs.
 -   Added Async Mode to the "Ticks" tab ([#57](https://github.com/8Crafter-Studios/Bedrock-World-Editor/issues/57)).
 -   Added Async Mode to the "Ticking Areas" tab ([#57](https://github.com/8Crafter-Studios/Bedrock-World-Editor/issues/57)).
+-   Added Async Mode to the "Structures" tab ([#57](https://github.com/8Crafter-Studios/Bedrock-World-Editor/issues/57)).
 
 ## Changes
 
@@ -112,7 +113,6 @@
 -   Fixed a bug where the error dialogs for opening files with the app would display the path as `[object Object]` instead of the actual path.
 -   Fixed a bug where the contents of entries with a data type of `NBT` or `custom/JSONNBT` could not be searched with the `contents` query in the "View Files" tab.
 -   Fixed a bug where if a search in the "View Files" tab that did not include an `nbt` or `contents` query was performed prior to performing a search with an `nbt` or `contents` query, without switching off of the tab in between the two searches, the NBT data for the entries would not be loaded, resulting in the `nbt` or `contents` query not matching any NBT search results.
--   The "Ticking Areas" tab now no longer unnecessarily loads the dynamic properties data.
 
 ## Performance Improvements
 
@@ -120,6 +120,8 @@
 -   Minor performance improvements.
 -   Major load time improvements to the "View Files" tab, as it no longer unnecessarily reads and parses the data of all entries with a data type of `ascii` or `int` during the initial load.
 -   Improved the loading times of the "Ticks" tab by removing an unnecessary sort operation on the list of pending tick keys.
+-   The "Ticking Areas" tab now no longer unnecessarily loads the dynamic properties data.
+-   The "Structures" tab now no longer unnecessarily loads the dynamic properties data.
 
 # v1.0.0-beta.34
 
@@ -156,8 +158,11 @@
 
 -   Fixed small seams that could sometimes appear in bewteen chunks on the 2D world map.
 -   The "Replace Image" option of the map editor now correctly clears the existing map image before drawing the selected image onto the map.
--   The "Maps" tab now no longer unnecessarily loads the dynamic properties data.
 -   The zoom buttons on the 2D world map now keep the map centered on the same position.
+
+## Performance Improvements
+
+-   The "Maps" tab now no longer unnecessarily loads the dynamic properties data.
 
 # v1.0.0-beta.33